requirements.txt
pyttsx
selenium
tkinter
SpeechRecognition
pyaudio
gtts
playaudio